requirements.txt
Django==2.2.3
django-fsm==2.6.1
django-fsm-admin==1.2.4
djangorestframework==3.10.0
gunicorn==19.9.0
pytz==2019.1
sqlparse==0.3.0
